Make functions take metadata directly from netcdf file
Previously, functions that find or insert modelmeta objects based on metadata from a netcdf file did so by delegating to a function that accepts generic (non-netcdf) params for the metadata. This introduces complexity and potential for error, though some tests are easier to set up. Decided it was better to invest effort in test harness and simplify code under test. Now the delegation functions are removed, and all find/insert functions accept a single netcdf parameter.