Rewire how index actions are determined
R script and earlier draft of this code had a somewhat confusing set of logic to determine how a given input file should be treated. This code is based on an identification of the various ways an input file can relate to the database (new file, previously indexed file, copy of previously indexed file, symlink to previously indexed file, etc.), with a single, unnested condition for each case. This makes the code simpler to read and debug.