Allow index_netcdf to reference relative paths
A common use case for indexing a set of files is to cd into the local
directory containing the files and then to run "index_netcdf *". This
will fail, unfortunately, because the database tracks absolute paths
and index_netcdf naively accepts the paths from the command line.

The PR tells index_netcdf to use the normalized absolutized version of
each path, so that the command line args will match up with the paths
in the database.