Remove memoization from index_netcdf.py
The memoization code in index_netcdf.py was both problematic and
ineffective. As part of the keys to the cache, it used open NetCDF
file objects and as part of the values of the cache were Netcdf
variable objects (which also contained references to the NetCDF
files). On exit, the main branch of the code would close and clean up
the files, leaving everything in the cache stranged. When the cache
got cleaned up, it would leave a long string of error messages:

Exception ignored in: 'netCDF4._netCDF4.Dataset.__dealloc__'
AttributeError: __getattribute__
AttributeError: __getattribute__

for each stranded object. While this didn't necessarily cause any real
errors at execution time, it was messy.

Furthermore, the expensive part of indexing is I/O: going out to read
data from the disk and going out to the database. The memoization, as
written, didn't actually mitigate this. It only cached objects,
referencing the file. I measured the differences in the test suite
between memoized and non-memoized and it was on the order of
miliseconds (noise). I recommend simply removing this code, as this
patch does.